Revert "UNCHECKED CHANGES: Added facility to configure the CSV-Export the same way adding it at the v8 version of the module. Only difference: reconstruction of the configuration file (in CmsCsvExportBean.java) is adjusted. Details: - Added CSV-Export configuration facility - Renamed CmsCvsExportBean to CmsCsvExportBean and adjusted calls - Refactored CmsCsvExportBean"
This reverts commit 9a18a24e466cfaaa63d7b7217dd3d44c067d09fb.
