+++ com.alkacon.opencms.formgenerator/src/com/alkacon/opencms/formgenerator/database/export/messages.properties
@@ -1,3 +1,2 @@
 LOG_WARN_EXPORT_DATEFORMAT_ILLEGAL_2 = Wrong date format configured in module parameter {0}: {1}. See http://java.sun.com/j2se/1.5.0/docs/api/java/text/SimpleDateFormat.html for information on date format to configure.
-LOG_WARN_EXPORT_NONEXISTING_COLUMNS_CONFIGURED_0 = At least one of the columns configured for CSV export does not exist. Please adjust the form configuration.
-LOG_WARN_EXPORT_RESOURCE_NOT_PRESENT_1 = Could not find the containerpage where the form is placed. Thus, a possible column configuration is not used. ErrorMessage: {0} 
+
